refactor: using latest node features and rename NaiveBayes to GaussianNB
