+++ .babelrc
@@ -1,0 +1,3 @@
+{
+  "plugins": ["transform-es2015-modules-commonjs"]
+}
