+++ .travis.yml
@@ -1,6 +1,5 @@
 language: node_js
 node_js:
-  - "4"
-  - "6"
-  - "stable"
+  - 8
+  - lts/boron #v6
 sudo: false
