+++ rollup.config.js
@@ -1,0 +1,5 @@
+export default {
+    entry: 'src/index.js',
+    format: 'cjs',
+    dest: 'lib/index.js'
+};
