--- test/.eslintrc.yml
@@ -1,2 +1,0 @@
-env:
-  mocha: true
